+++ pygsp/tests/test_all.py
@@ -13,12 +13,14 @@ from pygsp.tests import test_filters
 from pygsp.tests import test_utils
 from pygsp.tests import test_docstrings
 from pygsp.tests import test_plotting
+from pygsp.tests import test_learning
 
 
 suites = []
 suites.append(test_graphs.suite)
 suites.append(test_filters.suite)
 suites.append(test_utils.suite)
+suites.append(test_learning.suite)
 suites.append(test_docstrings.suite)
 suites.append(test_plotting.suite)  # TODO: can SIGSEGV if not last
 suite = unittest.TestSuite(suites)